# ev-charging-operations/contracts/openapi/auth-v1.yaml
openapi: 3.1.0
info:
  title: EVCO 认证服务 API
  version: 1.3.0
  description: |
    平台认证服务的 OpenAPI 契约。
    覆盖登录（图形验证码）、登出、刷新、撤销、M1 微信登录 Mock、
    IOT 协同模式 SSO 免登录（一次性 ticket 换会话）与 profile（权限/菜单/部署模式）。
    引用 platform-foundation-v1.yaml 的公共组件。
    v1.2.0：全部 200 响应统一补齐 ApiResponse 信封（与实现一致）；
    m1/wechat-mock 标注交付批次（依赖 evco_customer personal_user 表）。
paths:
  /api/v1/auth/captcha:
    get:
      summary: 获取图形验证码
      description: 登录页调用；返回 captchaId 与 Base64 PNG 图片，Redis 存储 120 秒，校验一次性使用。
      operationId: getCaptcha
      responses:
        '200':
          description: 验证码已生成
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './platform-foundation-v1.yaml#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/CaptchaResponse'
  /api/v1/auth/login:
    post:
      summary: 平台管理端登录
      operationId: login
      parameters:
        - $ref: './platform-foundation-v1.yaml#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './platform-foundation-v1.yaml#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 认证失败（用户名/密码错误、验证码错误、账号锁定）
  /api/v1/auth/sso/tickets:
    post:
      summary: IOT 推送 SSO 一次性 ticket（服务间接口）
      description: |
        仅 IOT 协同模式启用。IOT 后端在用户点击「充电运营平台」入口时调用，
        携带服务间凭证（X-API-Key / mTLS），推送一次性 ticket 与 IOT 用户标识；
        P1 侧 Redis 存储（键 sso:ticket:{ticket}，SET ... EX 120 NX 唯一性写入，
        撞号返回 VALIDATION_ERROR），供随后 /sso/login 换取会话。
      operationId: pushSsoTicket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SsoTicketPushRequest'
      responses:
        '200':
          description: ticket 已接收
          content:
            application/json:
              schema:
                $ref: './platform-foundation-v1.yaml#/components/schemas/ApiResponse'
        '400':
          description: ticket 撞号（键已存在，NX 写入失败）
        '401':
          description: 服务间凭证无效
        '403':
          description: 非IOT协同模式，该入口关闭
  /api/v1/auth/sso/login:
    post:
      summary: SSO 免登录（ticket 换会话）
      description: |
        仅 IOT 协同模式启用（独立部署模式入口整体关闭，调用一律 403）。用户从 IOT 平台跳转 P1 前端 /sso/login?ticket=xxx 后调用；
        校验使用 Redis GETDEL 原子读取并删除（并发同 ticket 有且仅有一个成功，防重放），
        读到 nil 统一返回 SSO_TICKET_INVALID（不区分不存在/已用/已过期，防探测）；
        消费成功后按 iotUserId（即 iam_user 主键，IOT 推送用户沿用 IOT ID 作主键）定位
        iam_user（source=IOT_PUSH、status=active），
        签发与密码登录一致的令牌对；仅识别已推送用户，本地不存在或不可用返回 SSO_USER_NOT_FOUND
        （ticket 保持已消费，不回滚复活）。Redis 故障时整体 fail-closed。
      operationId: ssoLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SsoLoginRequest'
      responses:
        '200':
          description: SSO 登录成功，直接进入平台
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './platform-foundation-v1.yaml#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '401':
          description: ticket 无效、已使用或已过期（SSO_TICKET_INVALID）
        '403':
          description: 本地无此用户或用户不可用（SSO_USER_NOT_FOUND）
  /api/v1/auth/profile:
    get:
      summary: 当前用户档案（权限、菜单树与部署模式）
      description: |
        登录后调用。返回用户基础信息、全部权限码（resource:read/write）、
        deployment_mode（IOT_COLLABORATIVE/STANDALONE）与后端推导的可见菜单树。
        菜单可见性规则：叶子菜单（页面）可见=拥有对应 resource:read；
        父级菜单可见=其下存在任意可见叶子，全部不可见则父级整棵隐藏（不出现空菜单）。
      operationId: getProfile
      responses:
        '200':
          description: 当前用户档案
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './platform-foundation-v1.yaml#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProfileResponse'
        '401':
          description: 未登录或令牌无效
  /api/v1/auth/logout:
    post:
      summary: 主动登出
      operationId: logout
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: 登出成功
          content:
            application/json:
              schema:
                $ref: './platform-foundation-v1.yaml#/components/schemas/ApiResponse'
  /api/v1/auth/refresh:
    post:
      summary: 刷新 access_token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: 刷新成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './platform-foundation-v1.yaml#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
        '401':
          description: refresh_token 已使用或已撤销
  /api/v1/auth/revoke:
    post:
      summary: 被动撤销（管理员强制下线）
      operationId: revokeSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RevokeRequest'
      responses:
        '200':
          description: 撤销成功
          content:
            application/json:
              schema:
                $ref: './platform-foundation-v1.yaml#/components/schemas/ApiResponse'
  /api/v1/auth/m1/wechat-mock:
    post:
      summary: M1 微信登录 Mock（C1-Mock）
      description: |
        W2-D-12 交付物。依赖 evco_customer personal_user/channel_identity 表
        （设计包 §3.2 迁移批次）；当前批次（evco_iam §3.1）未实现，随个人身份批次交付。
        C1-Mock 通过不等于 C1-W 真实门禁通过。
      operationId: m1WechatMockLogin
      parameters:
        - $ref: './platform-foundation-v1.yaml#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/M1WechatMockLoginRequest'
      responses:
        '200':
          description: Mock 登录成功
          content:
            application/json:
              schema:
                allOf:
                  - $ref: './platform-foundation-v1.yaml#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/LoginResponse'
components:
  schemas:
    CaptchaResponse:
      type: object
      required: [captchaId, imageBase64, expiresIn]
      properties:
        captchaId:
          type: string
          description: 验证码标识；登录时回传。
        imageBase64:
          type: string
          description: Base64 PNG 图形验证码（4-6 位字母数字，大小写不敏感）。
        expiresIn:
          type: integer
          description: 有效秒数（120）。
    LoginRequest:
      type: object
      additionalProperties: false
      required: [username, password, captchaId, captchaCode]
      properties:
        username:
          type: string
        password:
          type: string
          format: password
        captchaId:
          type: string
          description: 图形验证码标识；来自 GET /api/v1/auth/captcha。
        captchaCode:
          type: string
          description: 图形验证码内容；一次性校验，校验后立即失效。
        tenantId:
          type: string
          description: 租户 ID（雪花 ID 字符串形式）；平台管理员可省略。
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: 短 TTL 访问令牌。
        refreshToken:
          type: string
          description: 长 TTL 刷新令牌；一次性使用。
        expiresIn:
          type: integer
          description: access_token 过期秒数。
        tokenType:
          type: string
          default: Bearer
        userId:
          type: string
          description: 用户 ID（雪花 ID 字符串形式）。
        displayName:
          type: string
        roles:
          type: array
          items: { type: string }
        deploymentMode:
          type: string
          enum: [IOT_COLLABORATIVE, STANDALONE]
          description: 平台部署模式；前端据此控制 IOT 维护数据域写按钮显隐。
    SsoTicketPushRequest:
      type: object
      additionalProperties: false
      required: [ticket, iotUserId, expireAt]
      properties:
        ticket:
          type: string
          description: 一次性票据随机串（≥32 字符，CSPRNG 生成）；写入键 sso:ticket:{ticket}，SET EX 120 NX（撞号拒绝）；P1 侧校验用 GETDEL 原子消费；原文不写日志/审计（仅记 SHA-256 指纹前 8 位）。
        iotUserId:
          type: string
          description: IOT 侧用户 ID（雪花 ID 字符串形式）；IOT 协同模式推送用户沿用 IOT ID 作 iam_user 主键，换会话按主键定位本地用户（source=IOT_PUSH）。
        displayName:
          type: string
          description: IOT 侧展示名；仅用于审计记录，不覆盖本地用户资料。
        expireAt:
          type: string
          format: date-time
          description: 票据过期时间（UTC）。
    SsoLoginRequest:
      type: object
      additionalProperties: false
      required: [ticket]
      properties:
        ticket:
          type: string
          description: 跳转 URL 携带的一次性票据；校验通过后立即失效。
    ProfileResponse:
      type: object
      properties:
        userId:
          type: string
          description: 用户 ID（雪花 ID 字符串形式）。
        username:
          type: string
        displayName:
          type: string
        roles:
          type: array
          items: { type: string }
        permissions:
          type: array
          items: { type: string }
          description: 当前用户全部权限码（resource:read / resource:write）。
        deploymentMode:
          type: string
          enum: [IOT_COLLABORATIVE, STANDALONE]
        menus:
          type: array
          description: 后端按权限推导的可见菜单树；父级菜单仅在其下存在可见叶子时返回。
          items:
            $ref: '#/components/schemas/MenuNode'
    MenuNode:
      type: object
      required: [key, title]
      properties:
        key:
          type: string
          description: 菜单标识（kebab-case，如 system-config）。
        title:
          type: string
          description: 菜单中文名，与《充电运营平台菜单架构》一致。
        path:
          type: string
          description: 叶子菜单（页面）路由，如 /system-config/permission-management/platform-admin；父级菜单可为空。
        permission:
          type: string
          description: 叶子菜单入口权限码（resource:read）；父级菜单可为空。
        children:
          type: array
          items:
            $ref: '#/components/schemas/MenuNode'
    LogoutRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    RevokeRequest:
      type: object
      required: [sessionId]
      properties:
        sessionId:
          type: string
          description: 会话 ID（雪花 ID 字符串形式）。
        reason:
          type: string
    M1WechatMockLoginRequest:
      type: object
      additionalProperties: false
      required: [mockCode]
      properties:
        mockCode:
          type: string
          description: Mock 授权码；仅用于 C1-Mock 测试。
        applicationInstanceId:
          type: string